+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 <!-- Add new, unreleased items here. -->
+- Fix gitRepo.fetch() for when no branchName is given
+- Fix gitRepo.getHeadSha() & add tests
 
 ## v1.0.1 [10-05-2017]
 - Update package.json description